Restructured project and refactored cache implementation.
Specialized data structures now get their own package.
We will now only export interfaces.
Implementations will not be exported and will have functions for
initializing the structure as an implementation of some interface.
As cache is in its own package it now uses the List interface, as
implemented by linkedList.